fix: Make "None" option clickable in select dialog on formatter settings and also clean settings when selecting "None" (#1395)
The "None" option in the select dialog was unclickable because the click handler was prematurely returning when the item's value was falsy. This affected options with a `null` value, such as the "None" option for formatters.

This commit modifies the click handler to allow `null` values to be selected. The handler now only returns early if the item's value is `undefined`.

Also when selected `None` on formatters settings then delete the key saved in settings for that language, as it is not require in case of `null`